Add stride option to override the padding added by the driver
V4L2 allows specifying extra padding at the end of a line of pixels, the
driver can also override this to have a higher minimum padding than 0.

This adds a libdng_set_stride(dng, stride) function to override the
minimum padding that normally gets calculated by libdng when packing the
data.

The makedng utility now has a -S option to specify the stride in bytes
for a single row of pixels.